requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
opencv-python
numpy
Pillow
sqlalchemy==2.0.25
scikit-learn
